+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.9.0 - UNRELEASED
 * Ensure that `Filesystem::destroy` is always called
 * Remove request parameter to `Filesystem::destroy`
+* Fix `MountOption::AllowRoot`. Previously using it resulted in a crash.
 
 ## 0.8.0 - 2021-06-11
 * Deprecate `mount()`